+++ defaults.py
@@ -2,7 +2,7 @@ import os
 import stat
 import sys
 from copy import deepcopy
-from Common import LoudDict
+from appCommon.Common import LoudDict
 from camlib import to_dict, CNCjob, Geometry
 import simplejson
 import logging